requirements.txt
# Core Deep Learning
torch>=2.1.0
pytorch-lightning>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face ecosystem
transformers>=4.36.0
datasets>=2.15.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# LoRA and Quantization
bitsandbytes>=0.41.0
peft>=0.7.0

# Configuration
hydra-core>=1.3.2
omegaconf>=2.3.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.0
pyarrow>=14.0.1
scikit-learn>=1.3.0
